requirements.txt
# ── Core ──────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pyjwt==2.9.0
pydantic==2.9.2
jinja2==3.1.4
python-multipart==0.0.12
python-dotenv==1.0.1

# ── Testing ───────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2      # also used by TestClient
